bpt-komdigi-chatbot-BE/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=2.6.4
pymongo==4.6.2
python-multipart==0.0.9
python-dotenv==1.0.1
qdrant-client==1.8.0
redis==5.0.8
pandas==2.2.2
openpyxl==3.1.5
docx2txt==0.8
easyocr==1.7.1
langchain-google-vertexai>=1.0.0
langchain-community>=0.2.0
langchain-qdrant
pdfplumber==0.11.0
pypdf
langchain-openai>=0.2.0
# Semantic clustering untuk top questions analytics (gratis, lokal, support Bahasa Indonesia)
sentence-transformers>=2.7.0
